feat: 상품 삭제시 id가 없으면 예외를 던지는 기능 추가

+++ src/main/java/cart/service/ProductService.java
@@ -1,14 +1,13 @@
 package cart.service;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.stereotype.Service;
-
 import cart.dao.ProductDao;
 import cart.dao.entity.Product;
 import cart.dto.ProductRequest;
 import cart.dto.ProductResponse;
+import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
+import org.springframework.stereotype.Service;
 
 @Service
 public class ProductService {
@@ -37,7 +36,15 @@ public class ProductService {
     }
 
     public void delete(final Long id) {
+        if (doesNotExist(id)) {
+            throw new NoSuchElementException("상품을 찾을 수 없습니다.");
+        }
+
         productDao.delete(id);
+    }
+
+    private boolean doesNotExist(final Long id) {
+        return !productDao.existBy(id);
     }
 
     public void update(final Long id, final ProductRequest request) {
